moved files to private, also added default message to let uiser know there are no notifications or if its laoding
